Improved unit test.

+++ tests/ConnectionTest.php
@@ -1,15 +1,19 @@
 <?php
 
-include("../core/register_user.php");
-include("../core/connect_user.php");
+include(realpath(dirname(__FILE__) .
+        "/../core/register_user.php"));
+include(realpath(dirname(__FILE__) .
+    "/../core/connect_user.php"));
 use PHPUnit\Framework\TestCase;
 
 class ConnectionTest extends TestCase
 {
     public function testFindUser()
     {
-        $file = "../data/testIndex";
-        unlink($file);
+        $file = realpath(dirname(__FILE__) . "/../data/") .
+            "/indexTest";
+        if (file_exists($file))
+            unlink($file);
         $mail = "edouard@gmail.com";
         addEntryToAccountIndex($mail, $file);
         $this->assertTrue(checkIfUserExists($mail, $file));
@@ -17,8 +21,10 @@ class ConnectionTest extends TestCase
 
     public function testDontFindUser()
     {
-        $file = "../data/testIndex";
-        unlink($file);
+        $file = realpath(dirname(__FILE__) . "/../data/") .
+        "/indexTest";
+        if (file_exists($file))
+            unlink($file);
         $mail = "edouard@gmail.com";
         addEntryToAccountIndex($mail, $file);
         $this->assertFalse(checkIfUserExists("prout", $file));
